kubeflow/ingest/.ipynb_checkpoints/requirements-checkpoint.txt
pathlib2
requests
wget
azure-storage-file-datalake